fixed bug.

+++ txlcn-tm/src/main/java/com/codingapi/txlcn/tm/support/restapi/vo/ExceptionInfo.java
@@ -53,7 +53,7 @@ public class ExceptionInfo {
     /**
      * 异常情况。-1 【未知】 0 【TxManager通知事务】， 1 【TxClient查询事务状态】 2 【事务发起方通知事务组】
      */
-    private int registrar;
+    private short registrar;
 
     /**
      * 异常状态 0 待处理 1已处理
